fixed compile errors due to Integer->int cast

+++ cydonia/src/main/java/de/findus/cydonia/server/GameServer.java
@@ -328,11 +328,11 @@ public class GameServer extends MainController{
 		if(target != null && other != null) {
 			if(other.getName().startsWith("player")) {
 				Player p = getPlayerController().getPlayer(Integer.parseInt(other.getName().substring(6)));
-				if(p.getTeam() == (int) target.getUserData("team")) { // own target
+				if(p.getTeam() == ((Integer)target.getUserData("team")).intValue()) { // own target
 					if(p.getFlag() != null) {
 						int stolenflagid = p.getFlag().getId();
 						p.setFlag(null);
-						Flag f = getWorldController().getFlag((int)target.getUserData("id"));
+						Flag f = getWorldController().getFlag(((Integer)target.getUserData("id")).intValue());
 						if(f.isInBase()) {
 							System.out.println("Team " + p.getTeam() + " scored");
 							FlagEvent event = new FlagEvent(FlagEvent.SCORE, p.getId(), stolenflagid, true);
@@ -341,7 +341,7 @@ public class GameServer extends MainController{
 					}
 				}else { // opponents target
 					if(p.getFlag() == null) {
-						Flag f = getWorldController().getFlag((int)target.getUserData("id"));
+						Flag f = getWorldController().getFlag(((Integer)target.getUserData("id")).intValue());
 						if(f.isInBase()) {
 							f.setInBase(false);
 							System.out.println("Team " + p.getTeam() + " took flag");
